2 lv - 수정

+++ src/main/java/calculator/AddOperator.java
@@ -3,7 +3,7 @@ package calculator;
 public class AddOperator extends ArithmeticCalculator {
 
   @Override
-  public double operate(int firstNumber, int secondNumber) {
+  public double operate(Double firstNumber, Double secondNumber) {
     return firstNumber + secondNumber;
   }
 }
