+++ src/main/java/calculator/ArithmeticCalculator.java
@@ -2,5 +2,6 @@ package calculator;
 
 public abstract class ArithmeticCalculator {
 
-  public abstract double operate(int firstNumber, int secondNumber);
+  public abstract double operate(Double firstNumber, Double secondNumber);
+
 }
