+++ src/main/java/calculator/DivideOperator.java
@@ -3,7 +3,7 @@ package calculator;
 public class DivideOperator extends ArithmeticCalculator{
 
   @Override
-  public double operate(int firstNumber, int secondNumber) {
+  public double operate(Double firstNumber, Double secondNumber) {
     if (secondNumber == 0) {
       throw new ArithmeticException("Divide by zero");
     }
